# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Dish:
    properties:
      cookingTime:
        type: integer
      cost:
        type: number
      description:
        type: string
      image:
        type: string
      name:
        type: string
      restaurantId:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    required:
    - cookingTime
    - cost
    - name
    - restaurantId
    - status
    - type
    - weight
    type: object
  domain.DishesCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  domain.GetAllDishes:
    properties:
      cost:
        type: number
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
  domain.GetDishByID:
    properties:
      cookingTime:
        type: integer
      cost:
        type: number
      description:
        type: string
      image:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
  domain.GetDishesByRestaurant:
    properties:
      category:
        type: string
      categoryId:
        type: string
      dishes:
        items:
          $ref: '#/definitions/domain.GetAllDishes'
        type: array
    type: object
...
  domain.UpdateDish:
    properties:
      cookingTime:
        type: integer
      cost:
        type: number
      description:
        type: string
      image:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
...
  handler.dataResponse:
    properties:
      data: {}
    type: object
  handler.getAllDishesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GetAllDishes'
        type: array
    type: object
...
  handler.getDishResponse:
    properties:
      data:
        $ref: '#/definitions/domain.GetDishByID'
    type: object
  handler.getDishesByRestaurantResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GetDishesByRestaurant'
        type: array
    type: object
  handler.getDishesTypes:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.DishesCategory'
        type: array
    type: object
...
  handler.response:
    properties:
      message:
        type: string
    type: object
  handler.uploadResponse:
    properties:
      url:
        type: string
    type: object
host: 165.232.68.67:8080
info:
  contact: {}
  description: REST API for Restaurant-Assistant service
  title: Restaurant-Assistant API
  version: "1.0"
paths:
...
  /dish/:
    post:
      consumes:
      - application/json
      description: create dish
      parameters:
      - description: create dish info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Dish'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
  /dish/{id}:
    delete:
      consumes:
      - application/json
      description: delete dish
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
    get:
      consumes:
      - application/json
      description: get dish information
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getDishResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
    put:
      consumes:
      - application/json
      description: update dish
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      - description: order update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateDish'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getDishResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
  /dish/all:
    get:
      consumes:
      - application/json
      description: Get all dishes
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getAllDishesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
  /dish/image/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Endpoint to upload and update images
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: dish, restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - image
  /dish/restaurant/{id}:
    get:
      consumes:
      - application/json
      description: Get all restaurant dishes
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getAllDishesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
  /dish/types:
    get:
      consumes:
      - application/json
      description: Get all types of dishes
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getDishesTypes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
...
  /restaurant/dishes/{id}:
    get:
      consumes:
      - application/json
      description: Get all restaurant dishes with category
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getDishesByRestaurantResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - dish
  /restaurant/image/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Endpoint to upload and update images
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: dish, restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - image
...
securityDefinitions:
  ManagerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"